update README to correct page size bytes vs words
Hi,

SPM_PAGESIZE is in bytes. Maybe it used to be in words, but now it is in bytes. You can see that here:

http://www.atmel.com/webdoc/avrlibcreferencemanual/group__avr__io.html

Furthermore, your examples have flash page sizes of 128 words = 256 bytes, not 128 bytes, as written here. And your Python code defaults to 128 words. This wording mistake in the README threw me off for a while because I was working with an ATMega 328PB and that uses a 128 byte = 64 word page size. 

For backwards compatibility, you probably need to leave everything like so and not change your Python code. But I think the real fix, since SPM_PAGESIZE is in bytes, is to just do everything in bytes in your Python options and your description here. It's needlessly confusing, perhaps for historical reasons.